Replace InstallerMode with AllMode.
Since the old implementation did fetch both updates and all
packages in one go and keept the updates in a seperate list,
we had to use AllMode to get the packages of both lists. Now
the implementation fetches updates and keeps it in one list
and all packages (including updates) in another one. So AllMode
is now InstallerMode && UpdaterMode.