Fix accidental partly/ full uninstall in package manager.
In package manager mode a component (one to replace) might
be scheduled for uninstall, but we don't know if it should
be uninstalled. To figure this out we need to check the
actual replacement if it is still checked. If so, skip the
undo operation. This avoids an update if we actually do a
completely different component install/ uninstall.