Properly fix the use of the binary file raw pointer.
Make sure we pass the shared pointer around instead of the
actual raw pointer, this makes sure it get's destroyed at
the right point in time. As long as we run the installer,
the binary stays read opened during the whole application
run, while running the updater/ package manager/ uninstaller
will close the binary as soon as we mapped all data.