Keep the updater application as class member.
I can't see any reason to be able to set it from the outside,
thus it makes no sense to have that functionality at all.
To be able to work properly the app had to be set, but could
be easily forgotten, leading to a crash later on. Re-parsing
should be removed as well, will do it next.