Make sure we can use possible to replace installed components.
Means, once we notice that a component replaces a other,
we will remove it from the passed components list and
keep it inside the replacement hash. This needs to be
taken into account when we resolve component dependencies,
as a possible update might be a replacement, but not
scheduled for install while an other update to install
can have a dependency of the component to replace.

Change-Id: I00c0ad025f6a3640cfecfb8e3a1f36c6ec8cd6d8
Reviewed-on: http://codereview.qt.nokia.com/4250
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Tim Jenssen <tim.jenssen@nokia.com>
Reviewed-by: Niels Weber <niels.2.weber@nokia.com>
Reviewed-by: Karsten Heimrich <karsten.heimrich@nokia.com>