Fixes QTIFW-41.
If we have written a new binary and running as updater, check
the state the core engine is in to decide if we should restart.
Since the ui "Finish" button sets the core status to Canceled,
avoid to restart. In case we have written a new maintenance
tool, the core engine knows the need to restart, take this into
account on the ui restart page and "finish" the dialog.

Change-Id: I40fbc2d352d16f606e7fdc4d5bc61ff9f48bd56c
Reviewed-on: http://codereview.qt-project.org/5626
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Niels Weber <niels.2.weber@nokia.com>
Reviewed-by: Tim Jenssen <tim.jenssen@nokia.com>
Reviewed-by: Karsten Heimrich <karsten.heimrich@nokia.com>