Make sure to create components for local installed packages.
Should fix a problem in case we point the repo to a complete
different url or a component get's removed from the original
server. The problem then was, there was no component created
for the local installed package, which in case of update or
package managment would remove the local package as no component
cold be found for the stored operations. -> see runPackageUpdater

Change-Id: I7394691c533609526d09e994f509b47b9cb9c206
Reviewed-by: Tim Jenssen <tim.jenssen@nokia.com>