Fix crash while deleting components.
To delete it's childs, the component uses qDeleteAll on the childs list,
but once a child get's deleted it removes itself from the calling
components child list and then messes up the list iterators. The child
list passed to qDeleteAll needs to be a deep copy of the list the
component owns to not mess with the original the calliing component holds.

Change-Id: I638ffbb576a9543b7197c8bf267239f31197e274
Reviewed-by: Karsten Heimrich <karsten.heimrich@nokia.com>