Add missing braces. Fix broken meta download finished.
In case the meta data download finished, but there was no
hash given, the current package name wasn't reset so on the
next run we would try to download the same file again and
ended up in an endless loop. Reset the current name also in
case there's no has to fix this loop.

Change-Id: I1cdaa6e1fbbf13e3bb1852dc3d3fd8fffa28ee63
Reviewed-by: Niels Weber <niels.2.weber@nokia.com>
Reviewed-by: Tim Jenssen <tim.jenssen@nokia.com>