Compress all content, except already compressed files.
Task-number: QTIFW-136

This is now needed as we put the version number in front of
all content inside the data folder. This however results in
unexpected file name changes once the files are put inside the
repository and later are installed on the target system. If
we keep all files in zips, we can version them and install it
with the original name.

Change-Id: Ia700386c3f77de33818395a07d418c5ec40a0a14
Reviewed-by: Oliver Wolff <oliver.wolff@digia.com>
Reviewed-by: Niels Weber <niels.weber@digia.com>
Reviewed-by: Tim Jenssen <tim.jenssen@digia.com>