Update uninstaller creation on Mac to work with separate installer data
Installer data is stored in a separate file on Mac, and writeUninstaller
must be updated to understand that fact. To simplify code, the
uninstaller magic marker is also stored in a separate data file.

Also fixed writeUninstallerBinary to write the operations start and end
values that are part of the data header.

Change-Id: I3ee5d1eaa0542d6187a587fb72165c3a5b7306ec
Reviewed-by: Karsten Heimrich <karsten.heimrich@digia.com>