Rewrote file type registering.
Done now as described in MSDN. Make sure we honor the all users
flag. Reset the file type only if there has nothing changed
inbetween. Remove the ProgID entirely. Add the registered ProgId
inside the common "Open with" menu as well.

If passed a ProgId, use this one as key inside the classes hive.
If not given, fallback to the more hacky *_auto_file naming scheme.

Change-Id: I3f1c1a2301ad19b7fe215b65a8dd9c9419022f6c
Reviewed-by: Niels Weber <niels.weber@digia.com>
Reviewed-by: Tim Jenssen <tim.jenssen@digia.com>
Reviewed-by: Karsten Heimrich <karsten.heimrich@digia.com>