Fix a possible race and some wrong assumptions about file ownership.
The download archive job does not need to cleanup the files afterwards,
that's taken care of by the meta data download. That one will provide
the infrastructure where the files are downloaded at. Also prevent the
meta job from removing the files once they are fetched in case of a
cancel event (which does not entirely mean we are going to shutdown).

Change-Id: I66eeff30ef4cabb485dd4f300b2917deb7557867
Reviewed-by: Kai Koehne <kai.koehne@digia.com>
Reviewed-by: Niels Weber <niels.weber@digia.com>