Fix endless loop.
Caused by the fact that the settings wrapper did not support
anything different then native format, we had to trick the wrapper
into using its default QSettings object which in turn uses a
QFile (which roundtrips to the admin server) to write the settings
out (behind the scenes). The blocking appeared only on Linux cause
there we try a native call fcntl(handle, F_SETLKW, &fl) to lock the
file during sync, which blocks endless caused of the missing rights.
The fix is to use the settings wrapper also for ini format, as both
are supported (Ini and Native).

Change-Id: I73131c4adf85ba175ba6af1e18acccc29451b14f
Reviewed-by: Kai Koehne <kai.koehne@digia.com>