Fix possible case of uninstalling whole windows.
Task-number: QTIFW-511

In case we couldn't read the .dat file or the config file,
target dir will return the path the application was started.
Now we bail out early if one of the files is missing. Fixes
also some outdated code mess when we switched from a single
binary maintenance tool to the split binary version.

Change-Id: I4c70ac4ca63142873ed1521df47d74331669b576
Reviewed-by: Kai Koehne <kai.koehne@digia.com>
Reviewed-by: Niels Weber <niels.weber@digia.com>