Do not download update.xml files if they are on disk.
We did download the file from server in a different code path (put to
tmp), hand over the so called update source info to again download the
file from tmp to tmp to finally parse it. So if the scheme is "file" or
"resource", avoid this step and set the file directly.

Change-Id: Idaf8867142f7c2eb9c644321f458bd51373e98e6
Reviewed-by: Kai Koehne <kai.koehne@digia.com>