Fix broken update option.
We need to read the marker on our own, BinaryContent::binaryLayout()
throws cause it cannot read the full layout in case of maintenance
tool binary (part of the stuff is inside the .dat file).
To be able to read the layout from the .dat file, we need to open it.
Pass the right file to relace, in both cases the executable.

Change-Id: Ic76e78a2ee289f3d59d51cc8984d84ab5edf5b46
Reviewed-by: Kai Koehne <kai.koehne@digia.com>