Add --framework-version argument
This prints the version number of the Qt Installer Framework. In contrast
to the more verbose '--version' argument it can be used safely by
other applications to check for the version.

Also, we might at one point decide to give the IFW user control of the
--version output: He might want to run his own version scheme to also
take into account the packaged content.

Change-Id: Ia609310659ac45c5acffe318ecb6480eeb419ee8
Reviewed-by: Jarek Kobus <jaroslaw.kobus@digia.com>
Reviewed-by: Karsten Heimrich <karsten.heimrich@digia.com>