Fix uninstall after installing into a target with elevated permissions.
We wrongly did check the binary for write permissions, which we
actually had. Still we would bring the elevation dialog up and request
permissions, though only because the mkdir operation had to perform
its undo elevated as it created the installation directory. On Windows
the maintenance tool is locked until it is shutdown, so the operation did
not perform a full undo and the permissions got dropped again.
Then the VB script (which -does- the actual delete) was setup and launched
but without elevation and thus failed also to remove the maintenance tool.

Task-number: QTIFW-479
Task-number: QTIFW-447
Change-Id: I08e9757b34713f73472acde5d1b0f6e140c601e2
Reviewed-by: Niels Weber <niels.weber@digia.com>