Fix crash experienced while closing the application.
Signal the keep alive object that we are going to shutdown.
This needs to be done because the parent object is a global
static and the order of destruction is unknown.
The revised code of KeepAliveObject::start() method fixes
one of several warnings ("QSocketNotifier: socket notifiers
cannot be {enabled/disabled} from another thread").

Change-Id: I68e2d507e697a5d94bdc73f016601affc9913846
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>