Don't set modified state without real modification
setApplicationName and setApplicationVersion probably
should not set the modified state to true as they are called
way before actual installation is started. In case of real
changes happening, modified will always be set to true by the
other changes anyway.

Without this change, an existing installation in the default
installation directory of the new installer might be destroyed
when aborting the new installer. This should probably be investigated
further so that nothing is ever written before the real install
is started.

Task-number: QTIFW-583
Change-Id: I9921a33134133ae28433c4bfb726fe79ba93c20e
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>