Fix online installation into directory with elevated permissions.
Inside the Resource class we can't use QFile, as that will access
the elevated running server to read from a only client side mapped
file. Adjust Resource::readData to behave properly with the file
engine's nativeRead(...) implementation, e.g. on Unix it returns
-1 if called to read with a size of 0 bytes.
Do not write to the socket if we are running in a different thread
than the one the socket was created in. Fix logically broken connect
function.

Task-number: QTIFW-572
Change-Id: I993b5d3a5c217b0aedbbc27837dce2619e51d224
Reviewed-by: Niels Weber <niels.weber@theqtcompany.com>
Reviewed-by: Takayuki ORITO <iori.ayane@gmail.com>
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>