Clean up CreateDesktopEntry operation
We can't delay the deletion of the file if we then try to replace it
with the backup file/write to it. In addition, since we're on Unix,
the deletion of the file can only fail if we don't have sufficient rights,
and delaying the deletion until a 'reboot' doesn't fix this.

Also, don't treat any errors in the undo step as errors that should show
up in the UI: There are lots of reasons people might clean up their
settings.

Change-Id: I2e85f94e4fa427e0911ffab3a0111e4fb058ff1a
Reviewed-by: Niels Weber <niels.weber@theqtcompany.com>