Fix corruption of uninstaller operations after 'soft restart'
After updates have been installed, the user can restart the maintenance
tool GUI, and check for updates again. If no updates are found then,
availableComponents() will remain empty, letting fetchUpdaterPackages() nuke
also all 'updaterComponentsDeps'. This in turn breaks
sortOperationsBasedOnComponentDependencies(), which needs the list of
components to calculate the order.

Work around this by forcing a write of the uninstaller data before the
core is re-initialized. This let the soft updater behave like an
independent updater run ... if no updates are installed, the uninstaller
is not written.

Task-number: QTIFW-600
Change-Id: I83659491b8e2c0c7a0478362a6c3a4dfaec542b3
Reviewed-by: Tim Jenssen <tim.jenssen@theqtcompany.com>