Remove country from most translations
If we provide an en_us translation only this won't be picked up
by users having configured 'en', or 'en_uk'. This also mirrors the
naming of the qt translations.

The exception is zh_CN though, because the difference between
simplified and traditional characters is big. I'm not sure whether
taiwanese people would rather prefer traditional characters or english.
For Qt we have both zh_CN and zh_TW, maybe we will get a zh_TW in the
future too...

Change-Id: I73490337aef8f54da14b6feca3a035e1e405d65b
Reviewed-by: Niels Weber <niels.weber@theqtcompany.com>