Delete components only after they have been removed from the lists
This fixes a crash when a component destructor e.g. tries to unregister
a widget: This might lead to the creation of a new scripting engine,
which then tries to register all components into the engine, including
components that have already been deleted ...

Avoid this by first cleaning the lists, and then doing the actual
deletion.

Task-number: QTIFW-612
Change-Id: I708a1ade3b3cc15006eefe394c2d81bf47ebe754
Reviewed-by: Niels Weber <niels.weber@theqtcompany.com>