Do not defer deletion of socket in RemoteObject
The 7z backend can create and close a considerable number of RemoteObjects
without ever returning to the event loop. For big archives the socket
objects are therefore piling up, slowing down the extraction over time.

QTcpSocket::close is documented to close the connection automatically in its
destructor, so deleting it instantly is safe.

Change-Id: I95323b53e6886a7bc6c0b59dd3c13ade4aa9567c
Reviewed-by: Niels Weber <niels.weber@theqtcompany.com>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@theqtcompany.com>