RemoteServer: Make sure we don't write to stderr
In adminauthorization_x11.cpp we redirect stderr of the child
process to a pipe, but do not read from it anymore after sudo
completed. Which means, if the server process continues to write
into the pipe it'll block at some point.

Avoid this by always closing the stderr handle in production mode.

Change-Id: If1e9578f12d7a0f60811f9c2b43cd9c18adfea59
Reviewed-by: Niels Weber <niels.weber@theqtcompany.com>
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>