Fix update from 1.6 and earlier to 2.0 installerbase.
The update was broken caused by the attribute rename. We already
print a warning and read the old value, though we did not set it
for the new introduced tags in case of an update.
Thus we where not able to read the .ini file and the values stored,
but did fallback to the use the initial values from config.xml.

Task-number: QTIFW-625
Change-Id: Iad1c96c690de81c5653864a3777c78bd43007249
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>
Reviewed-by: Niels Weber <niels.weber@theqtcompany.com>
Reviewed-by: Iikka Eklund <iikka.eklund@theqtcompany.com>