Restore original behavior for component.installationRequested
component.installationRequested() should return true also for
components that are selected through dependencies (after they
have been resolved). This restores the original behavior on 1.6.

To keep the change minimal the 'new' behavior has been saved
in a method isSelectedForInstallation . This should be further
refactored in the master branch.

Task-number: QTBUG-633
Change-Id: I2cb936a399927252a4ad4dd81f73683fa3c28a01
Reviewed-by: Jarek Kobus <jaroslaw.kobus@theqtcompany.com>