Fix embedding of installbase manifest
Redefining QMAKE_MANIFEST is currently broken in Qt for newer compilers:
The content is not embedded at all if the toolchain supports "embedding".

But the hack is not needed anyway: Just disable the Qt manifest magic,
and include our custom manifest file via .rc. This way we can also avoid
changing the mkspec for embed_manifest_exe, embed_manifest_dll.

Change-Id: I5155ddabe17f0b61b2d7b9c5b9cd912ad6c53dfe
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>