Fix dubious warning about RequiresAdminRights tag.
The tag was added once component script is parsed and we did
know that addElevatedOperation is called from script. However,
the initial commit 7ccd0acdcf35b957e9e7d04cd81902bac362ad04 did
not append the node to the parent package update node. Now we
decided to have the option statically and declarative available,
so we can remove the warning.

Task-number: QTIFW-670
Change-Id: I8cd8f702391d7bf80fc4c42074e1fd6d2d4abc57
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>