Replace UpdateSourcesInfo and UpdateSourceInfo classes.
Introduce a new struct PackageInfo, which is similar to the removed
UpdateSourceInfo struct. The new struct name reflects the actual use
better, as we deal with package here (either packages to install or
updates to apply) and not updates only. Also remove the container
class UpdateSourcesInfo, we can simple reuse a existing Qt container
here. Adjust all occurrences that use/ used the classes.

Change-Id: I526e24cbf1664f0ab0ad18153f60c2c10b6909d8
Reviewed-by: Jarek Kobus <jaroslaw.kobus@theqtcompany.com>