Replace waitForBytesWritten with flush
Unfortunately waitForByesWritten is known to fail from time to time:
QTBUG-24451 . We introduced it for making sure large data is actually
send, but flush() seems to serve the same purporse. The only drawback
is that flush() operates on the concrete socket, and therefore needs
to be called on the caller's side.

Change-Id: Ic92db298b3dad2c3f9eaa47d160aec517e33e284
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>