OS X: Make sure real user uid == euid in elevated mode
This fixes a bug that prevents changing existing .ini files
during an elevated installation.

QFileSystemEngine::fillMetaData() on Unix uses :access() to check
whether the user can write to a file, based on the real user id.
Because OS X's AuthorizationExecuteWithPrivileges() does keep the
original real user id also in the elevated process,
QFileInfo::isWritable() was returning false for existing files,
which in turn let QSaveFile, and therefore QSettings, to not even
attempt to write to the file.

On Linux, we use sudo to start the elevated process, which already
resets the real user id by default.

Task-number: QTIFW-709
Change-Id: I5bfbd631f579412045b242a08baf206a5d444500
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>