Fix missing operation 'installer' value and update documentation.
Commit 478a8a1db3eed2 was done under the assumption that after
UpdateOperation::fromXml(...) existing values are still set.
That is wrong and we have to set the 'installer' value again.

The same commit was trying to fix the empty 'installer' value
in the operations destructor, though the value got removed in
PackageManagerCore::writeMaintenanceToolBinaryData() already.
We now skip it once we write out the XML.

Change-Id: Ia029ae4583862f4cbc6feb1c757604bf7b042a8f
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>