Fix: .dat file gets deleted after multiple component changes.
We can close the .dat file in maintenance mode. There is no need
to keep it open since no packages are in there and resources are
mapped into memory anyway. Fixes commit 7f2c98c8 that updates the
.dat file on soft restart. On Windows updating did not work cause
the .dat is opened and locked by the OS. To overcome locked files
we run a VB script that does a deferred rename once the lock is
gone, but in case of a soft restart we now did start several of
them depending on the amount of restarts.
That had the undesired effect of several replace/rename operations,
of which only 2 "succeeded":

- First script: Remove .dat | Rename .dat.new -> .dat
- Second script: Remove .dat | no .dat anymore...

Task-number: QTIFW-689
Change-Id: Ic3ee1b418890eabe5b854dc7879cd7bc118f9240
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>