Fix violated assertion in error handling of binarycreator
The errorMessageFrom7zResult() function asserts (via a qFatal) that the
given status code is indeed an error, i.e. not S_OK. However, the
condition 'res != S_OK || !tempFile.exists()' permitted that the
function is called with S_OK (in case !tempFile.exists() is true),
violating an assertion.

This patch adjusts the error handling such that the details of the error
returned by errorMessageFrom7zResult() are only included in the
exception message if the status code is not S_OK.

Change-Id: I15d686cd35c918093ec90c56a7bcc3f0e6638cfb
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>