Automatically chose to perform uninstall if appropriate
This patch makes the maintenance tool automatically chose (but not
start) deinstallation if there are no valid repositories available.
Without this, all three options are available but the first two will
show an error message if there is not even one repository available.

This simplifies the maintenance tool for purely offline installers which
use no repositories whatsoever.

Change-Id: If496ac6858191c758a7bae6ec4a9434f88ddaac0
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>