Fixed running binarycreator if the temporary directory name contains spaces
The binarycreator program fails to generate an archive if the temporary
directory name contains spaces. This is not uncommon on Windows since
the temporary directory is beneath the home directory, e.g.
C:\Users\<username>\AppData\Local\Temp. If the user name contains spaces
(as it does in my case, the user name is "Frerich Raabe") binarycreator
fails with

  [64] Warning: QFile::remove: Empty or null file name (C:\Qt\MSVC12\5.5.0-src\qtbase\src\corelib\io\qfile.cpp:498, bool __thiscall QFile::remove(void))
  Caught exception: Cannot create archive "C:\Users\Frerich": internal code: E_FAIL

This was caused by createArchive() assembling a command line without
escaping either the 'target' or any of the 'sources' values.

Instead of adding escaping only to split the command line again, let's
drop the approach of constructing a command and then splitting it
completely. Instead, let's build an array of arguments right away.

Change-Id: I284c1b5a27e9edd3717243ea7979149ab2033d64
Task-number: QTIFW-787
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>