Fixed occasional crash on Windows when terminating installer
The static RemoteClient object (defined in RemoteClient::instance())
which gets destroyed at the very end of the process lifetime would -- as
part of its constructor -- execute the RemoteClientPrivate destructor.

The RemoteClientPrivate destructor did non-trivial work though, trying
to shut down the connection which involves accessing the LocalServer
instance. Executing this code at the very end of the process lifetime
would sometimes cause a crash on Windows with a back trace into the
QMutex guts.

Let's fix this by getting more control over the destruction order; the
RemoteClient instance is no longer a static object but allocated
dynamically. Callers can explicitly call 'destroy' to dispose of the
instance as part of their shutdown routine.

Change-Id: Ie6773ccc0276ad6285919e57e9ed190bab02be3f
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>