Factored flushing of VerboseWriter into dedicated method
This patch introduces a (yet unused) VerboseWriter::flush() method which
can be used to explicitly trigger flushing the installation log. This
permits clients to decide when to do the flushing instead of having to
rely on the destructor being called.

The destructor will call flush() if needed - whether flushing is needed
is defined in terms of the 'perFileBuffer': if it's open, flushing is
needed - otherwise, no flushing is needed.

Change-Id: I17f1df96e5ddf3b6f613d647c28b90b6973c9393
Reviewed-by: Katja Marttila <katja.marttila@theqtcompany.com>