Fix timeout errors while building app bundles files on OS X
...by simply not using any timeout anymore. I believe this was the
original intention of the code (it's arguably hideous that the default
value of the waitForFinished() method is a 30000ms timeout).

What happened for me is that the mkdmg.sh script took longer than 30
seconds. As a consequence

  1. the 'waitForFinished()' call returned
  2. The mkdgm.sh script was deleted
  3. The QProcess object went out of scope, terminating the shell process
  4. The binarycreator process terminated successfully

The resulting .dmg file seemed plausible (it was fairly large already)
but it wasn't usable yet since the hdiutil process started by mkdmg.sh
was still running, i.e. the .dmg file was still growing. Hence, the
caller of binarycreator (e.g. some CI system) continued processing a
malformed .dmg file (and the temporary file created by mkdmg.sh never
got deleted).

Last but not least, proper error handling would have been nice of course
but I considered that to be out of scope for this particular commit.

Change-Id: I767ff89d20802306189660804cea617c148a943e
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>
Reviewed-by: Katja Marttila <katja.marttila@theqtcompany.com>