Add alternative option where to read the installer-framework sha1
The "--version" uses the sha1 of the git checkout of the
installer-framework.

This works when making the build from git checkout.

For Coin (continuous integration) the installer-framework builds are
made from stored src packages where the .git directory is not present.

In this case read the sha1 from the .tag instead which is present in
the Coin produced src packages.

Change-Id: Ic6387c844699f4410c043a4df84727bb5c1108ae
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@theqtcompany.com>
Reviewed-by: Simon Hausmann <simon.hausmann@theqtcompany.com>