Fixed deleting files when uninstalling on OS X
Uninstalling packages on OS X didn't seem to work very reliably, it
often (always?) seemed to not remove a couple of files. When clicking
the 'Show Progress' button while uninstalling, it became apparent that
the paths to all the files to remove were computed wrongly: instead of
deleting e.g. the file

  /tmp/installdir/foo.txt

The uninstaller attempted to remove

  /tmp/installdir/Uninstall.app/foo.txt

What happens is when uninstalling, the ExtractArchiveOperation class
(which performs deletion of installed files in its 'undoOperation'
reimplementation) reads the files which were installed from the XML
markup embedded in the maintenance tool binary (this happens in
UpdateOperation::fromXml()). The paths to the files which were installed
make use of a special '@RELOCATABLE_PATH@' placeholder which denotes the
effective installation directory. This placeholder is meant to be
replaced with the actual installation directory, which is assumed to be
the directory where the maintenance tool resides, i.e. what
QCoreApplication::applicationDirPath();.

On OS X however, the maintenance tool is an app bundle, i.e. the actual
binary is not in

  /tmp/installdir/Uninstall

but rather

  /tmp/installdir/Uninstall.app/Contents/MacOS/Uninstall

The code already called the QInstaller:isInBundle() utility function to
handle this case, but did so wrongly: isInBundle() returns the directory
to the app bundle, i.e.

  /tmp/installdir/Uninstall.app

What we want though is just '/tmp/installdir'. Hence, let's go one
directory up.

Change-Id: I927a453d7fbdd1048ff3d2172bd3cfec3a563b4a
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>