Resize banner image to fit default installer width
When the Modern wizard style is used, the size of the banner picture
limits the width of the installer and that's a problem on high DPI
displays. Scaling and resizing of the UI does not work correctly in that
case. Also the WizardDefaultWidth option from config.xml is ignored
completely.

Let's try to solve this by resizing the banner image to fit the
installers default width. This is technically not backwards compatible
(existing installers may suddenly appear to be much more narrow or wide
than before) but makes the installer work much better on high DPI
displays in which case the fonts tends to be very big (in terms of
pixels) and the banner image looks very small.

If the old behavior is needed, setting the 'WizardDefaultWidth' value to
be equal to the width of the banner image will do the job.

Change-Id: Ib311df3bbc277c328e015c1095ae4d35663c42e1
Reviewed-by: Niels Weber <niels.weber@theqtcompany.com>
Reviewed-by: Katja Marttila <katja.marttila@theqtcompany.com>