Fixed final 'Finish' message on OS X
On OS X, the final page of the installer would say something like

  Click Done to exit the %1 wizard.

However, the final "commit" button is only actually labelled "Done" when
using the natie OS X style.

When specifying a different style in the config.xml file, e.g. 'Modern',
the commit button is labelled the same as on other operating systems:
"Finish". Hence, in such cases the final message would be wrong.

Fix this by not hardcoding any particular label in the message but
rather querying the default button text - this works with all styles.
This also required setting the text a little later (when entering the
page) because only then we can safely access the pointer returned by the
gui() method.

Change-Id: I285c6e376c8450457fcce20941240f9e29350de7
Reviewed-by: Niels Weber <niels.weber@theqtcompany.com>
Reviewed-by: Katja Marttila <katja.marttila@theqtcompany.com>