Fixed detecting problems while launching sudo
The installer would always consider the exit status of the sudo process,
but if executing sudo itself fails (e.g. because the file does not
exist, or because of insufficient permissions), a success exit status
(zero) was returned.

Fix this by rather returning the (negative) errno value in case anything
goes wrong with execve.

Change-Id: Iac5051e998982c6a31fa49cb6f4290cf66050c44
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@theqtcompany.com>
Reviewed-by: Simon Hausmann <simon.hausmann@theqtcompany.com>